refactor: eliminate SQLiteBackend shim, convert build() to async, explicit __init__ export
Unit 5: Redirect all 16 files importing SQLiteBackend from sqlite.py
to import directly from async_sqlite.py. Delete the __getattr__ lazy
re-export shim and remove SQLiteBackend from sqlite.py's __all__.

Unit 6: Convert 6 sync test methods in test_branching.py from .build()
(which used asyncio.run()) to async def + await .async_build(). Delete
the sync build() method from ConversationBuilder.

Unit 7: Replace lazy __getattr__ in polylogue/__init__.py with explicit
import of ConversationRepository, add to __all__.

Unit 4 (async_index rename) was reverted: the plan incorrectly assumed
no sync counterpart exists, but storage/index.py has sync FTS functions
used by health.py, search_messages(), and test_fts5.py.
