+++ tests/integration/test_security.py
@@ -31,7 +31,7 @@ from polylogue.sources.source import (
     MAX_UNCOMPRESSED_SIZE,
     iter_source_conversations,
 )
-from polylogue.storage.backends.sqlite import SQLiteBackend
+from polylogue.storage.backends.async_sqlite import SQLiteBackend
 from polylogue.storage.repository import ConversationRepository
 from polylogue.storage.search import escape_fts5_query
 from polylogue.storage.store import AttachmentRecord, ConversationRecord
