+++ tests/unit/storage/test_search_misc.py
@@ -30,10 +30,8 @@ from polylogue.storage.backends.schema import (
     _migrate_v8_to_v9,
     _migrate_v9_to_v10,
 )
-from polylogue.storage.backends.sqlite import (
-    SQLiteBackend,
-    connection_context,
-)
+from polylogue.storage.backends.async_sqlite import SQLiteBackend
+from polylogue.storage.backends.sqlite import connection_context
 from polylogue.storage.store import ConversationRecord
