+++ tests/unit/storage/test_sqlite_ops.py
@@ -4,10 +4,8 @@ from __future__ import annotations
 
 import pytest
 
-from polylogue.storage.backends.sqlite import (
-    SQLiteBackend,
-    open_connection,
-)
+from polylogue.storage.backends.async_sqlite import SQLiteBackend
+from polylogue.storage.backends.sqlite import open_connection
 from polylogue.storage.store import (
     AttachmentRecord,
     ConversationRecord,
